requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
pydantic-settings
PyJWT
pydantic[email]
python-multipart
requests
websockets